requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-rest-swagger==2.2.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.13
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.12
Pillow==5.3.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
pytz==2018.7
PyYAML==3.13
requests==2.21.0
simplejson==3.16.0
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
